Various small code fixes. Some unwanted code removed.
git-svn-id: https://swig.svn.sourceforge.net/svnroot/swig/branches/gsoc2012-doxygen@13551 626c5289-ae23-0410-ae9c-e8d60b6d4f22

+++ Source/DoxygenTranslator/src/TokenList.cpp
@@ -45,7 +45,7 @@ TokenList TokenList::tokenizeDoxygenComment(const std::string &doxygenComment, c
         // check if it's one of the '!!!', '***', '///' of any length
         char c = currentWord[0];
         isPlainString = false;
-        for (int i=0; i<currentWord.size(); i++)
+        for (size_t i=0; i<currentWord.size(); i++)
           if (currentWord[i] != c) {
             isPlainString = true;
             break;
